+++ controllers/identification.js
@@ -1,0 +1,49 @@
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const User = require('../models/user');
+const BadRequestError = require('../errors/BadRequestError');
+const ConflictError = require('../errors/ConflictError');
+const AuthenticationError = require('../errors/AuthenticationError');
+
+const { NODE_ENV, JWT_SECRET } = process.env;
+module.exports.createUser = (req, res, next) => {
+  const { email, password } = req.body;
+  bcrypt
+    .hash(password, 10)
+    .then((hash) => User.create({ email, password: hash }))
+    .then((user) => res.status(201).send({
+      email: user.email,
+      name: user.name,
+      about: user.about,
+      avatar: user.avatar,
+      _id: user._id,
+    }))
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        throw new BadRequestError('The email and password are required');
+      } else {
+        throw new ConflictError('User with this email already exist');
+      }
+    })
+    .catch(next);
+};
+
+module.exports.login = (req, res, next) => {
+  const { email, password } = req.body;
+  return User.findUserByCredentials(email, password)
+    .then((user) => {
+      if (user) {
+        const token = jwt.sign(
+          { _id: user._id },
+          NODE_ENV === 'production' ? JWT_SECRET : 'dev-secret',
+          { expiresIn: '7d' },
+        );
+        res.send({ token });
+      }
+    }).catch((err) => {
+      if (err) {
+        throw new AuthenticationError('authorization failed');
+      }
+    })
+    .catch(next);
+};
