+++ controllers/users.js
@@ -1,12 +1,5 @@
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
 const User = require('../models/user');
 const NotFoundError = require('../errors/NotFoundError');
-const BadRequestError = require('../errors/BadRequestError');
-const ConflictError = require('../errors/ConflictError');
-const AuthenticationError = require('../errors/AuthenticationError');
-
-const { NODE_ENV, JWT_SECRET } = process.env;
 
 module.exports.getCurrentUser = ((req, res, next) => {
   User.findById(req.user._id)
@@ -14,49 +7,8 @@ module.exports.getCurrentUser = ((req, res, next) => {
       if (user) {
         res.send({ data: user });
       } else {
-        throw new NotFoundError('the user was not found');
+        throw new NotFoundError('The user was not found');
       }
     })
     .catch(next);
 });
-module.exports.createUser = (req, res, next) => {
-  const { email, password } = req.body;
-  bcrypt
-    .hash(password, 10)
-    .then((hash) => User.create({ email, password: hash }))
-    .then((user) => res.status(201).send({
-      email: user.email,
-      name: user.name,
-      about: user.about,
-      avatar: user.avatar,
-      _id: user._id,
-    }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        throw new BadRequestError('The email and password are required');
-      } else {
-        throw new ConflictError('User with this email already exist');
-      }
-    })
-    .catch(next);
-};
-
-module.exports.login = (req, res, next) => {
-  const { email, password } = req.body;
-  return User.findUserByCredentials(email, password)
-    .then((user) => {
-      if (user) {
-        const token = jwt.sign(
-          { _id: user._id },
-          NODE_ENV === 'production' ? JWT_SECRET : 'dev-secret',
-          { expiresIn: '7d' },
-        );
-        res.send({ token });
-      }
-    }).catch((err) => {
-      if (err) {
-        throw new AuthenticationError('authorization failed');
-      }
-    })
-    .catch(next);
-};
