+++ routes/index.js
@@ -1,14 +1,11 @@
 const router = require('express').Router();
-const auth = require('../middleware/auth');
-const { createUser, login } = require('../controllers/users');
-const { validateUser } = require('../middleware/validations');
 
-const usersRouter = require('./articles');
+const identification = require('./identification');
+const usersRouter = require('./users');
 const articlesRouter = require('./articles');
 
-router.use('/', auth, usersRouter);
-router.use('/', auth, articlesRouter);
-router.use('/signin', validateUser, login);
-router.use('/signup', validateUser, createUser);
+router.use('/', identification);
+router.use('/', usersRouter);
+router.use('/', articlesRouter);
 
 module.exports = router;
